+++ src/components/header/header.scss
@@ -49,6 +49,16 @@ $link-actif: #88ADB7; // Couleur des liens actifs
   }
 }
 
+
+.header .user-firstname {
+  background-color: transparent; /* Assurez-vous que le fond est transparent */
+}
+
+.header .user-firstname:focus,
+.header .user-firstname:active {
+  background-color: transparent; /* Désactivez le fond bleu de focus ou active */
+}
+
 // Alignement central des liens dans l'en-tête
 .header-links {
   display: flex;
