+++ src/components/longinSigninModale/longinSigninModale.scss
@@ -1,0 +1,131 @@
+/* Variables */
+:root {
+  --bg-modal: rgba(0, 0, 0, 0.7);
+  --bg-modal-content: #262E33;
+  --bg-button: #228B22;
+  --bg-hover-button: #1d791d;
+  --bg-register-button: #007BFF; /* Nouvelle couleur pour le bouton "S'inscrire" */
+  --bg-hover-register-button: #022a55; /* Couleur du bouton au survol */
+  --color-title: #000000;
+  --color-text: #333;
+  --color-error: #e74c3c;
+}
+
+// Titre de la modale
+h2 {
+  font-size: 24px;
+  font-weight: 600;
+  color: white;
+  margin-bottom: 20px;
+  text-align: center;
+}
+/* Styles de la modal */
+.modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: var(--bg-modal);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 9999;
+  
+}
+
+.modal-content {
+  background-color: var(--bg-modal-content);
+  border: solid 1px white;
+  padding: 20px;
+  border-radius: 8px;
+  width: 400px;
+  position: relative;
+}
+
+/* Bouton de fermeture */
+.close {
+  position: absolute;
+  top: 0px;
+  right: 15px;
+  font-size: 30px;
+  cursor: pointer;
+  color: white;
+  transition: color 0.3s ease;
+  z-index: 3000;
+}
+
+
+.close:hover {
+  color: var(--color-error);
+}
+
+/* Switch entre le formulaire de connexion et d'inscription */
+.switch-form {
+  text-align: center;
+ 
+  margin-bottom: 1rem;
+
+ p {
+  margin-bottom: 10px;
+  margin-top: 10px;
+ }
+}
+
+.switch-form button {
+  background-color: var(--bg-button);
+  color: white;
+  padding: 10px 15px;
+ 
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.switch-form button:hover {
+  background-color: var(--bg-hover-button);
+}
+
+/* Styles pour le formulaire d'inscription */
+.register-options {
+  margin-top: 20px;
+}
+
+.register-options label {
+  display: block;
+  margin: 10px 0;
+  font-size: 16px;
+}
+
+.register-options input[type="radio"] {
+  margin-right: 10px;
+}
+
+button:disabled {
+  background-color: #0f0e0e;
+  cursor: not-allowed;
+}
+
+/* Styles pour le bouton "S'inscrire" */
+.register-button {
+  background-color: var(--bg-register-button); /* Utilisation de la nouvelle variable */
+  color: rgb(255, 255, 255);
+  width: 100%;
+  padding: 14px;
+  color: #fff;
+  font-size: 16px;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: background-color 0.3s ease, transform 0.2s ease;
+}
+
+.register-button:hover {
+  background-color: var(--bg-hover-register-button); /* Utilisation de la nouvelle variable pour le survol */
+}
+
+button:disabled {
+  background-color: #0f0e0e;
+  cursor: not-allowed;
+}
