--- src/components/modalLogin/modalLogin.scss
@@ -1,151 +1,0 @@
-:root {
-  --bg-modal: #262E33;
-  --bg-button: #228B22;
-  --bg-hover-button: #1d791d;
-  --color-title: #000000;
-}
-
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-
-  &-content {
-    background-color: var(--bg-modal);
-    border: 1px solid #ccc;
-    padding: 20px;
-    border-radius: 10px;
-    width: 90%;
-    max-width: 400px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    animation: fadeIn 0.3s ease-in-out;
-    position: relative;
-  }
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(-20px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.close {
-  position: absolute;
-  top: 10px;
-  right: 15px;
-  font-size: 20px;
-  font-weight: bold;
-  color: #888;
-  cursor: pointer;
-
-  &:hover {
-    color: #000;
-  }
-}
-
-h2 {
-  margin-bottom: 20px;
-  font-size: 24px;
-  text-align: center;
-  color: #fffafa;
-}
-
-.auth-form {
-  input {
-    display: block;
-    width: 90%;
-    margin: 10px;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    font-size: 16px;
-  }
-
-  button {
-    width: 100%;
-    padding: 10px;
-    background-color: var(--bg-button);
-    color: #fff;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-
-    &:hover {
-      background-color: var(--bg-hover-button);
-    }
-  }
-}
-
-.register-options {
-  p {
-    margin-bottom: 15px;
-    text-align: center;
-    color: #747474;
-    text-decoration: underline;
-  }
-
-  label {
-    display: block;
-    margin: 10px 0;
-    font-size: 16px;
-    cursor: pointer;
-  }
-
-  input {
-    margin-right: 10px;
-  }
-}
-
-.register-option-btn {
-  width: 100%;
-  padding: 10px;
-  margin: 10px 0;
-  background-color: var(--bg-button);
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  font-size: 16px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  }
-
- 
-
-
-.switch-form {
-  margin-top: 20px;
-  text-align: center;
-
-  p {
-    font-size: 14px;
-    color: #c0bebe;
-  }
-
-  button {
-    background: none;
-    color: #ffffff;
-    border: none;
-    font-size: 15px;
-    cursor: pointer;
-    text-decoration: underline;
-    padding: 0;
-
-    &:hover {
-      color: #4cbb31;
-    }
-  }
-}
